pyproject.toml
[tool.poetry]
name = "resume-ai-backend"
version = "0.1.0"
description = "Resume intelligence backend with OCR, RAG, and audit logging."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
license = "MIT"
packages = [{ include = "resume_ai", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
langchain = "^0.2.15"
langchain-openai = "^0.1.21"
langchain-community = "^0.2.15"
openai = "^1.44.0"
python-multipart = "^0.0.9"
motor = "^3.5.1"
qdrant-client = "^1.9.1"
paddleocr = "^2.7.3"
pymupdf = "^1.24.10"
pillow = "^10.4.0"
tenacity = "^8.4.2"
structlog = "^24.1.0"
python-ulid = "^2.7.0"
httpx = "^0.27.2"
pydantic-settings = "^2.4.0"
types-requests = "^2.32.0.20240712"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
pytest-mock = "^3.14.0"
coverage = "^7.6.2"
ruff = "^0.5.7"
mypy = "^1.11.1"
types-redis = "^4.6.0"
freezegun = "^1.5.0"

[tool.poetry.scripts]
resume-ai-cli = "resume_ai.interfaces.cli.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "B", "W", "C4", "N"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["resume_ai"]
profile = "black"

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_return_any = true
strict_optional = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = [
    "tests",
]

